user jwt

+++ controllers/user.controller.js
@@ -1,16 +1,17 @@
 import mongoose from 'mongoose';
-import { User } from '../models/user.model.js';
-import bcrypt from "bcryptjs"
+import { generateToken, User } from '../models/user.model.js';
+import bcrypt from "bcryptjs";
+import jwt from "jsonwebtoken";
 
 export async function signIn(req, res, next) {
     const { email, password } = req.body;
     const user = await User.findOne({ email }); //במקום email:email רק email כי זה באותו שם
     if (user) {
         const isSame = await bcrypt.compare(password, user.password);
-
         if (isSame) {
             user.password = "****";
-            return res.json(user);
+            const token = generateToken(user);
+            return res.json({user, token});
         }
         return next({ message: 'Auth Failed-user is not found', status: 401 })
     }
@@ -25,7 +26,8 @@ export async function signUp(req, res, next) {
         const user = new User({ name, email, password });
         await user.save();
         user.password = "****";
-        return res.status(201).json(user);
+        const token = generateToken(user);
+        return res.status(201).json({user, token});
     } catch (error) {
         return next({ message: error.message, status: 409 })
     }
@@ -36,12 +38,15 @@ export async function updateUser(req, res, next) {
     if (!mongoose.Types.ObjectId.isValid(id))
         return next({ message: 'id is not valid', status: 409 })
 
+    if (id !== req.user_id)
+        return next({message: 'user can update himself only', status: 403})
+
     try {
         let user = req.body;
         const { _id, name, email, password } = user; //unpacking
 
         if (id !== _id)
-            return next({ message: 'user can update himself', status: 409 })
+            return next({ message: 'there are 2 different id', status: 409 })
 
         user = await User.findByIdAndUpdate( //find user by id and update him
             _id,
@@ -62,7 +67,11 @@ export async function updateUser(req, res, next) {
 export async function deleteUser(req, res, next) {
     const id = req.params.id;
     if (!mongoose.Types.ObjectId.isValid(id))
-        next({ message: 'id is not valid' })
+        return next({ message: 'id is not valid' })
+
+    if (id !== req.user_id)
+        return next({message: 'user can delete himself only', status: 403})
+
     try {
         const user = await User.findByIdAndDelete(id);
         if (!user)
